Remove 2nd and 3rd TODO in static.tex
I wrote in the TODO's that it would take only 1 or 2 extra rules to
convert them to per-cmd rules mapped over a program, but I didn't
realise that we'd need to specify the mapping for each possible cmd
again, which is a large increase in rules. Therefore, I thought it best
to just leave it as-is.